% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvp.R
\name{bvp}
\alias{bvp}
\title{Read a Rmodel query}
\usage{
bvp(x, y = NULL, lambda = c(1, 1, 1), log = FALSE)
}
\arguments{
\item{filename}{Relative filepath to the current working directory. This must be (*.zip) or it will throw an error.}
}
\value{
A tibble from the inputted *.RData file.
}
\description{
read vectors to calculate bivariate poisson possibility and returns a vector.
}
\details{
This function read 1 to 2 vectors and then calculate bivariate poisson possibility and returns a vector.
}
\examples{
# this will stop
\dontrun{ bvp(FTHG~1, FTAG~1, lambda = c(lmb1, lmb2, lmb3)) }
#returns data as vector
 #setwd(system.file("data", package = "Rmodel"))
 #getwd()
 \dontrun{ bvp(FTHG~1, FTAG~1, lambda = c(lmb1, lmb2, lmb3)) }
}
\seealso{
\url{https://www.github.com/englianhu/Rmodel} for calculate the possibility of bivariate poisson model.

Other Rmodel functions: 
\code{\link{compileIndex2}()},
\code{\link{compileIndex}()}
}
\concept{Rmodel functions}
