% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileIndex2.R
\name{compileIndex2}
\alias{compileIndex2}
\title{Read a Rmodel query}
\usage{
compileIndex2(
  l1 = FTHG ~ 1,
  l2 = FTAG ~ 1,
  l1l2 = ~c(Home, Away) + c(Away, Home),
  l3 = ~1,
  data,
  maxit = 300,
  xi = NULL,
  fordate = NULL,
  fun = "glm",
  inflated = TRUE
)
}
\arguments{
\item{filename}{Relative filepath to the current working directory. This must be (*.zip) or it will throw an error.}
}
\value{
A list of dataset from the inputted *.RData file.
}
\description{
read vectors to calculate diagonal zero inflated bivariate poisson possibility and returns a list
}
\details{
This function read 1 to 2 vectors and then calculate bivariate poisson possibility and returns a vector.
}
\examples{
# this will stop
\dontrun{
compileIndex2(FTHG ~ 1, FTAG ~ 1, ~c(Home, Away) + c(Away, Home),
data = sample, xi = 0.0065, fordate = '2010-08-12', fun = 'speedglm') }
#returns data as vector
#setwd(system.file("data", package = "Rmodel"))
#getwd()
\dontrun{
compileIndex2(FTHG ~ 1, FTAG ~ 1, ~c(Home, Away) + c(Away, Home),
data = sample, xi = 0.0065, fordate = '2010-08-12') }
\dontrun{
em2008 = compileIndex2(FTHG~1, FTAG~1, ~c(Home, Away) + c(Away, Home), data = eng2008f) }
\dontrun{
em2009 = compileIndex2(FTHG~1, FTAG~1, ~c(Home, Away) + c(Away, Home), data = eng2009f) }
\dontrun{
em2010 = compileIndex2(FTHG~1, FTAG~1, ~c(Home, Away) + c(Away, Home), data = eng2010f) }
\dontrun{
em2011 = compileIndex2(FTHG~1, FTAG~1, ~c(Home, Away) + c(Away, Home), data = eng2011f) }
\dontrun{
em2012 = compileIndex2(FTHG~1, FTAG~1, ~c(Home, Away) + c(Away, Home), data = eng2012f) }
\dontrun{
em2013 = compileIndex2(FTHG~1, FTAG~1, ~c(Home, Away) + c(Away, Home), data = eng2013f) }
}
\seealso{
\url{https://www.github.com/englianhu/Rmodel} for calculate the possibility of bivariate poisson model.

Other Rmodel functions: 
\code{\link{bvp}()},
\code{\link{compileIndex}()}
}
\concept{Rmodel functions}
