% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmodel_formatprint.R
\name{formatPrintDec}
\alias{formatPrintDec}
\title{formatPrintDec}
\usage{
formatPrintDec(x, round = 4, big.mark = ".", decimal.mark = ",",
  scientific = FALSE)
}
\description{
formatPrintDec
}

