% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmodel_formatprint.R
\name{formatPrintNumber}
\alias{formatPrintNumber}
\title{formatPrintNumber}
\usage{
formatPrintNumber(x, round = 4, big.mark = ".", decimal.mark = ",",
  scientific = FALSE)
}
\description{
formatPrintNumber
}

