% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel_formatprint.R
\name{printPerformance}
\alias{printPerformance}
\title{printPerformance}
\usage{
printPerformance(listperformance, printConfMat = TRUE, printF1 = TRUE,
  printMetrics = TRUE, round = 5, big.mark = ".", decimal.mark = ",",
  scientific = FALSE)
}

