% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmodel_datamanipulation.R
\name{r.binning}
\alias{r.binning}
\title{r.binning}
\usage{
r.binning(var, div = NULL, ndiv = 10, labelText = FALSE,
  labelNumeric = FALSE, prefix = "Grup", link1 = " ", link2 = ", ")
}
\description{
r.binning
}
\seealso{
\code{\link{r.binning}} \code{\link{r.segment}} \code{\link{r.segment.target}} \code{\link{r.segment.dataframe}} \code{\link{r.segment.target.table}}
}

