% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmodel_modelevaluation.R
\name{r.classifierMetrics}
\alias{r.classifierMetrics}
\title{r.classifierMetrics}
\usage{
r.classifierMetrics(clustReal = NULL, clustModel = NULL, beta = 1,
  selectedClass = NULL, ...)
}
\description{
r.classifierMetrics
}

