% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmodel_clusters.R
\name{r.clusters.silhouette}
\alias{r.clusters.silhouette}
\title{r.clusters.silhouette}
\usage{
r.clusters.silhouette(data, clusters, dist = dist(data, "euclidean"))
}
\description{
r.clusters.silhouette
}

