% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel_dataanalysis.R
\name{r.data.missings}
\alias{r.data.missings}
\title{r.data.missings}
\usage{
r.data.missings(dades, outputDFNumeric = "dfMissingsNumeric",
  outputDFCategorical = "dfMissingsCategorical")
}
\description{
r.data.missings
}

