% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel_dataanalysis.R
\name{r.data.missingsCrossed}
\alias{r.data.missingsCrossed}
\title{r.data.missingsCrossed}
\usage{
r.data.missingsCrossed(dades, variablesFil, variablesCol,
  outputDF = "dfMissingsCrossed", outputDFPerc = "dfMissingsCrossedPerc",
  outputDFTotals = "dfMissingsCrossedTotals")
}
\description{
r.data.missingsCrossed
}

