% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmodel_modelevaluation.R
\name{r.gain.old}
\alias{r.gain.old}
\title{r.gain.old}
\usage{
r.gain.old(score, fuga, perc = 0.2, mode = 0)
}
\arguments{
\item{mode}{1 Optimista, -1 Pesimista, 0 Sense Modificar, 2 Promig, 3 Random}
}
\description{
r.gain.old
}

