% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmodel_modelevaluation.R
\name{r.gains}
\alias{r.gains}
\title{r.gains}
\usage{
r.gains(score, target, npoints = 20, mode = "def")
}
\arguments{
\item{mode}{\cr "def" = As it is
\cr "rnd" = Random
\cr "pos" = Optimist/Positive (Maximum)
\cr "neg" = Pessimist/Negative (Minimum)
\cr "avg" = Average of Optimist and Pessimist}
}
\description{
r.gains
}

