% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmodel.R
\name{r.getAdj}
\alias{r.getAdj}
\title{r.getAdj}
\usage{
r.getAdj(adj, quantileCutOff = 0.6, absolute = TRUE, normalize = TRUE,
  removeAutoCycles = TRUE, undirected = TRUE, useColumnNames = TRUE)
}
\description{
r.getAdj
}

