% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmodel.R
\name{r.getDF.cleaned}
\alias{r.getDF.cleaned}
\title{r.getDF.cleaned}
\usage{
r.getDF.cleaned(datosInput, filterNA = TRUE, filterNAN = TRUE,
  filterInfinite = TRUE, filterConstants = FALSE, verbose = FALSE)
}
\description{
r.getDF.cleaned
}

