% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmodel_modelevaluation.R
\name{r.performance.metrics}
\alias{r.performance.metrics}
\title{r.performance.metrics}
\usage{
r.performance.metrics(score, target, threshold = 0.5, beta = 1, round = 5,
  big.mark = ".", decimal.mark = ",", scientific = FALSE,
  printConfMat = TRUE, printF1 = TRUE, printMetrics = TRUE)
}
\description{
r.performance.metrics
}

