% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel_model_preprocess.R
\name{r.quantile_normalize}
\alias{r.quantile_normalize}
\title{r.quantile_normalize}
\usage{
r.quantile_normalize(var, nquantiles = 10, probs = seq(0, 1, by =
  1/nquantiles), normalize = TRUE, na.rm = TRUE, rightmost.closed = TRUE,
  all.inside = FALSE)
}

