% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel_model_preprocess.R
\name{r.quantileSplit}
\alias{r.quantileSplit}
\title{r.quantileSplit}
\usage{
r.quantileSplit(var, nquantiles = 10, probs = seq(0, 1, by = 1/nquantiles),
  na.rm = TRUE, rightmost.closed = TRUE, all.inside = FALSE,
  showSplits = FALSE)
}

