% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel_datamanipulation.R
\name{r.segment.target}
\alias{r.segment.target}
\title{r.segment.target}
\usage{
r.segment.target(dades, segmentacion, target, targetValue = NULL)
}
\arguments{
\item{dades}{dataframe}

\item{segmentacion}{Array of strings. Names of the columns to create segmentation. Those variables must be categorical.}

\item{target}{String. Name of the target column to calculate means for each segment. If targetValue is provided, then means are not calculated, and percentatge of counting targetValue is provided instead.}

\item{targetValue}{Numeric or Categorical. Value of target column to calculate counts instead of means.}
}
\value{
An array of percentages of target for each segmentation.
}
\description{
r.segment.target
}
\seealso{
\code{\link{r.binning}} \code{\link{r.segment}} \code{\link{r.segment.target}} \code{\link{r.segment.dataframe}} \code{\link{r.segment.target.table}}
}

