\name{r.data.missingsCrossed}
\alias{r.data.missingsCrossed}
\title{r.data.missingsCrossed}
\usage{
  r.data.missingsCrossed(dades, variablesFil, variablesCol,
    outputDF = "dfMissingsCrossed",
    outputDFPerc = "dfMissingsCrossedPerc",
    outputDFTotals = "dfMissingsCrossedTotals")
}
\description{
  r.data.missingsCrossed
}

