% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel_modelevaluation.R
\name{r.gain}
\alias{r.gain}
\title{r.gain}
\usage{
r.gain(score, target, perc = 0.2, mode = "def", target_value = NULL)
}
\arguments{
\item{mode}{\cr "def" = As it is
\cr "rnd" = Random
\cr "pos" = Optimist/Positive (Maximum)
\cr "neg" = Pessimist/Negative (Minimum)
\cr "avg" = Average of Optimist and Pessimist}
}
\description{
r.gain
}

